backend/requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Telegram API Client
telethon>=1.32.0

# Database
aiosqlite>=0.19.0

# Configuration
python-dotenv>=1.0.0

# Image Processing
Pillow>=10.0.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Utilities
pydantic>=2.4.0
pydantic-settings>=2.0.0
